feat: organize folders to DDD
