--- siaes/src/main/java/soat_fiap/siaes/api/vehicle/dto/CreateVehicleRequest.java
@@ -1,16 +1,0 @@
-package soat_fiap.siaes.api.vehicle.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import soat_fiap.siaes.domain.vehicle.Vehicle;
-
-public record CreateVehicleRequest(
-        @NotBlank String plate,
-        @NotBlank String brand,
-        @NotBlank String model,
-        @NotNull Integer year
-) {
-    public Vehicle toModel() {
-        return new Vehicle(plate, brand, model, year);
-    }
-}
