+++ siaes/src/main/java/soat_fiap/siaes/config/SecurityConfiguration.java
@@ -14,8 +14,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-import soat_fiap.siaes.service.AuthenticationService;
-import soat_fiap.siaes.domain.repository.UserRepository;
+import soat_fiap.siaes.shared.AuthenticationService;
+import soat_fiap.siaes.domain.user.repository.UserRepository;
 
 @Configuration
 @EnableWebSecurity
