--- siaes/src/main/java/soat_fiap/siaes/domain/enums/RoleEnum.java
@@ -1,22 +1,0 @@
-package soat_fiap.siaes.domain.enums;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum RoleEnum {
-    ADMIN("admin"),
-    COLLABORATOR("collaborator");
-
-    private final String role;
-
-    public static boolean contains(String role) {
-        for (RoleEnum r : RoleEnum.values()) {
-            if (r.getRole().equals(role)) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
