--- siaes/src/main/java/soat_fiap/siaes/domain/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package soat_fiap.siaes.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import soat_fiap.siaes.domain.entity.User;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, UUID> {
-    Optional<User> findByLogin(String login);
-}
