+++ siaes/src/main/java/soat_fiap/siaes/domain/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package soat_fiap.siaes.domain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import soat_fiap.siaes.domain.user.model.User;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, UUID> {
+    Optional<User> findByLogin(String login);
+}
