--- siaes/src/main/java/soat_fiap/siaes/infra/vehicle/VehicleRepository.java
@@ -1,11 +1,0 @@
-package soat_fiap.siaes.infra.vehicle;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import soat_fiap.siaes.domain.vehicle.Vehicle;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface VehicleRepository extends JpaRepository<Vehicle, UUID> {
-    boolean existsByPlate(String plate);
-}
