+++ siaes/src/main/java/soat_fiap/siaes/interfaces/user/dto/LoginRequest.java
@@ -1,0 +1,11 @@
+package soat_fiap.siaes.interfaces.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @NotBlank(message = "O login não pode estar vazio")
+        String login,
+
+        @NotBlank(message = "A senha não pode estar vazia")
+        String password) {
+}
