+++ siaes/src/main/java/soat_fiap/siaes/interfaces/user/dto/LoginResponse.java
@@ -1,0 +1,11 @@
+package soat_fiap.siaes.interfaces.user.dto;
+
+import soat_fiap.siaes.domain.user.model.RoleEnum;
+
+public record LoginResponse(
+        String accessToken,
+        String refreshToken,
+        String username,
+        RoleEnum roleEnum
+) {
+}
