+++ siaes/src/main/java/soat_fiap/siaes/interfaces/user/dto/MessageDTO.java
@@ -1,0 +1,6 @@
+package soat_fiap.siaes.interfaces.user.dto;
+
+public record MessageDTO(
+        String message
+) {
+}
