+++ siaes/src/main/java/soat_fiap/siaes/interfaces/user/dto/RefreshTokenDTO.java
@@ -1,0 +1,9 @@
+package soat_fiap.siaes.interfaces.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RefreshTokenDTO(
+        @NotBlank(message = "O refresh token não pode estar vazio")
+        String refreshToken
+) {
+}
