+++ siaes/src/main/java/soat_fiap/siaes/interfaces/vehicle/dto/UpdateVehicleRequest.java
@@ -1,0 +1,12 @@
+package soat_fiap.siaes.interfaces.vehicle.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record UpdateVehicleRequest(
+        @NotBlank String plate,
+        @NotBlank String brand,
+        @NotBlank String model,
+        @NotNull Integer year
+) {
+}
