+++ siaes/src/main/java/soat_fiap/siaes/interfaces/vehicle/dto/VehicleResponse.java
@@ -1,0 +1,15 @@
+package soat_fiap.siaes.interfaces.vehicle.dto;
+
+import soat_fiap.siaes.domain.vehicle.model.Vehicle;
+
+public record VehicleResponse(
+        String id,
+        String plate,
+        String brand,
+        String model,
+        int year
+) {
+    public VehicleResponse(Vehicle vehicle) {
+        this(vehicle.getIdAsString(), vehicle.getPlate(), vehicle.getBrand(), vehicle.getModel(), vehicle.getYear());
+    }
+}
