--- siaes/src/main/java/soat_fiap/siaes/rest/dto/LoginResponse.java
@@ -1,13 +1,0 @@
-package soat_fiap.siaes.rest.dto;
-
-import soat_fiap.siaes.domain.enums.RoleEnum;
-
-import java.util.UUID;
-
-public record LoginResponse(
-        String accessToken,
-        String refreshToken,
-        String username,
-        RoleEnum roleEnum
-) {
-}
