--- siaes/src/main/java/soat_fiap/siaes/rest/dto/MessageDTO.java
@@ -1,6 +1,0 @@
-package soat_fiap.siaes.rest.dto;
-
-public record MessageDTO(
-        String message
-) {
-}
