--- siaes/src/main/java/soat_fiap/siaes/rest/dto/RefreshTokenDTO.java
@@ -1,9 +1,0 @@
-package soat_fiap.siaes.rest.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RefreshTokenDTO(
-        @NotBlank(message = "O refresh token não pode estar vazio")
-        String refreshToken
-) {
-}
